REFACTOR: Move findRoutingRule to utility dir
